requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1.1
django-admin-thumbnails==0.2.6
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-session-timeout==0.1.0
idna==3.3
Pillow==8.3.2
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2021.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.1
tzdata==2022.1
urllib3==1.26.11
